Added improvements, e.g. better design, predefined filters, etc.

+++ src/imagesharpener.h
@@ -8,11 +8,11 @@ class ImageSharpener
 public:
     ImageSharpener();
     static QPixmap filterImage(const QPixmap *image);
-    static void setFilterArray(int x0_y0, int x1_y0, int x2_y0,
-                               int x0_y1, int x1_y1, int x2_y1,
-                               int x0_y2, int x1_y2, int x2_y2);
+    static void setFilterArray(float x0_y0, float x1_y0, float x2_y0,
+                               float x0_y1, float x1_y1, float x2_y1,
+                               float x0_y2, float x1_y2, float x2_y2);
 private:
-    static int filterArray[3][3];
+    static float filterArray[3][3];
 };
 
 #endif // IMAGESHARPENER_H
